Stub out initial Event graph component insertion Cypher
Adds in an initial Cypher query for bulk Event insertion. Required
doing some slight remodeling of the underlying chocolate-doom
Event data model, but all for the better. (Essentially the constraints
are spelled out in Event::toJson().)

This also required adding some quick and dirty `toMap()` methods to
all the consituent event classes since the Cypher Query can take
a Map<String, Object> a parameter value. Basically, the Query uses
almost a complete like-for-like representation of the Event record
that was originally received as JSON from a Doom instance.