Let the data flow! UDP plumbing is working.
Finally got what seems to be a viable config going for getting the data
to flow from in UDP listener to the Database with windowing going on in
between. (Question now is...do we really need to order the data?)

Still lacks the threading, but that's the next part. Graph components at
least being created around Events.

Also fixes some schema setup issues found along the way.

Minor change: bind to 127.0.0.1 and not 0.0.0.0...for some reason that
wasn't working on Windows, plus we shouldn't bind on external facing
adapters anyways without consent ;-)