Support configuring Groom from the CLI
Borrowed an approach from another of my projects, this implements
Argparse4j as a means of handling CLI args. It's a tad annoying in
some ways, but all configuration is now encapsulated in the Config
class and passed to a run() method for the app.

Some unit tests of two rough edges, but manually testing the rest.